New scene and renderer for LearnOpenGL demo

+++ OpenGLCourseApp/ImGuiWrapper.cpp
@@ -77,7 +77,7 @@ void ImGuiWrapper::End()
 {
 	// ImGui Rendering
 	ImGuiIO& io = ImGui::GetIO();
-	io.DisplaySize = ImVec2(m_Window->GetBufferWidth(), m_Window->GetBufferHeight());
+	io.DisplaySize = ImVec2((float)m_Window->GetBufferWidth(), (float)m_Window->GetBufferHeight());
 
 	// Rendering
 	ImGui::Render();
